feat(gpu): add image tags
Adds tags to the resulting image to identify if the image has gpu
support and the type and version of the support.

Tags added:

```json
{
  "gpu": "true",
  "gpu_nvidia_version": "470",
  "gpu_types": "nvidia"
}
```

The tags will always exis. If `gpu` is `"false"` then
`gpu_nvidia_version` and `gpu_types` will be the empty string `""`.

`gpu_types` is a comma specifed list. If in the future amd or opencl
types are added an image that supports both `amd` and `opencl` will be
tagged with:

```json
{
  "gpu_types": "amd,opencl"
}
```